Added a reset button and a newgrid button for user to set new dimensions, also added different logic on the createGrid function
I added inline styling to the container element, for both width and height, in the creatGrid function. I created a maxBoxSize in the global scope, and boxWidth and boxHeight in the function scope--and use them to add to the gridTemplateRows and gridTemplateColumns--so the box sizes adjust to the different dimenions. I also added min-height and min-width to the container-contain selector, so the grid never shrinks or increases. I added a reset button that removes the active class from the boxes, and new grid button that renders an innterHTML form when clicked. This form lets the user enter width and height to adjust the grids dimensions